pyproject.toml
[tool.poetry]
name = "pharmacy-information-system"
version = "0.1.0"
description = ""
authors = ["Kirill Kondrenko <kondrenkokp@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
mysql-connector-python = "^8.3.0"
mimesis = "^15.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
